[moveonly] Move MapIntoRange() to separate util/fastrange.h
